requirements.txt
SQLAlchemy==1.4.22
scp==0.13.6
Pillow==8.3.1
pyserial==3.5
PyYAML==5.4.1
requests==2.26.0
pytz==2021.1
pytest==6.2.4
